Move environment variable related codes into a new file.
